Will move FeedStore' to its own file
