+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudy/FeedCache/FeedStore.swift
@@ -1,0 +1,17 @@
+//
+//  FeedStore.swift
+//  ModularFeedTDDCaseStudy
+//
+//  Created by Clint M on 3/22/19.
+//  Copyright © 2019 Clint M. All rights reserved.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
